requirements.txt
flask
torch
opencv-python
pillow
tqdm
werkzeug
numpy
gunicorn